transport: add git-send-pack support to local/ssh. (#163)
* protocol/packp: add Packfile field to ReferenceUpdateRequest.
* protocol/packp: add NewReferenceUpdateRequestFromCapabilities.
* NewReferenceUpdateRequestFromCapabilities can be used to create
  a ReferenceUpdateRequest with initial capabilities compatible with
  the server.
* protocol/packp: fix new line handling on report status.
* transport/file: test error on unexisting command.

+++ plumbing/transport/file/common.go
@@ -69,6 +69,10 @@ func (c *command) StdoutPipe() (io.Reader, error) {
 
 // Close waits for the command to exit.
 func (c *command) Close() error {
+	if c.closed {
+		return nil
+	}
+
 	return c.cmd.Process.Kill()
 }
